+++ app/globals.css
@@ -74,3 +74,8 @@
     @apply bg-background text-foreground;
   }
 }
+
+input[type="time"]::-webkit-calendar-picker-indicator {
+  background: none;
+  display: none;
+}
